Add files via upload
I added two new scripts Zoom and SpawnUnit script. The zoom script enables the player to zoom in and out of the map. The player can get a better look at the units,buildings and map. The SpawnUnits script spawns new units next to the factory building in a specific position. Although only one type of unit spawns... for now.